+++ rotkehlchen/tasks/utils.py
@@ -21,7 +21,11 @@ log = RotkehlchenLogsAdapter(logger)
 
 def should_run_periodic_task(
         database: 'DBHandler',
-        key_name: Literal['last_data_updates_ts', 'last_evm_accounts_detect_ts'],
+        key_name: Literal[
+            'last_data_updates_ts',
+            'last_evm_accounts_detect_ts',
+            'last_spam_assets_detect_key',
+        ],
         refresh_period: int,
 ) -> bool:
     """