+++ docs/changelog.rst
@@ -2,8 +2,12 @@
 Changelog
 =========
 
+* :bug:`7009` PnL reports with an old type of event (such as DailyStats) following a swap will no longer fail with an exception.
+* :bug:`6998` If a username contains a '.' the user will now be able to log in properly again.
+* :feature:`6885` Users can now specify EVM chains for which no activity will be auto-detected by rotki.
 * :bug:`-` The welcome message at first login after a version upgrade will now have the correct link to the release notes.
 * :feature:`4841` The rotki data directory is now more organized.
+* :bug:`-` Creating, editing and deleting accounting rules will now update warnings when rendered events get affected in the history view.
 
 * :release:`1.31.0 <2023-11-24>`
 * :feature:`-` Oneinch v3 swaps should be supported in Ethereum mainnet.